requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2
drf-spectacular>=0.15.1,<0.16
Pillow
uwsgi>=2.0.19,<2.1